api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
google-cloud-bigquery>=3.0.0
pandas>=2.0.0
pymongo[srv]>=4.6.0
dnspython>=2.4.0
